Make BinpkgFetcher use longs instead of floats when synchronizing timestamps, since it's better not to rely on float comparison here.
svn path=/main/trunk/; revision=12302

+++ pym/_emerge/__init__.py
@@ -3535,12 +3535,12 @@ class BinpkgFetcher(SpawnProcess):
 				remote_mtime = bintree._remotepkgs[self.pkg.cpv].get("MTIME")
 				if remote_mtime is not None:
 					try:
-						remote_mtime = float(remote_mtime)
+						remote_mtime = long(remote_mtime)
 					except ValueError:
 						pass
 					else:
 						try:
-							local_mtime = os.stat(self.pkg_path).st_mtime
+							local_mtime = long(os.stat(self.pkg_path).st_mtime)
 						except OSError:
 							pass
 						else:
